// src/media/enums/file-type.enum.ts
export enum FileType {
  JPEG = 'image/jpeg',
  PNG = 'image/png',
  GIF = 'image/gif',
  SVG = 'image/svg+xml',
  
  MP4 = 'video/mp4',
  MPEG = 'video/mpeg',
  WEBM = 'video/webm',
  MSV = 'video/x-msvideo',
  
  PDF = 'application/pdf',
  TXT = 'text/plain',
  JSON = 'application/json',
  XLS = 'application/vnd.ms-excel',
  XLSX = 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
}